# Increase performance of lodash.sample