Increase performance by 30-120% depending if `null` is provided or not.

**bench.js**
```javascript
  .add('sample', () => { sample(['a', 'b', 'c']) })
  .add('sample null', () => { sample(null) })
  .add('sample []', () => { sample([]) })
  .add('lodash.sample', () => { _.sample(['a', 'b', 'c']) })
  .add('lodash.sample null', () => { _.sample(null) })
  .add('lodash.sample []', () => { _.sample([]) })
```
Benchmark results
```
sample x 60,896,314 ops/sec ±0.44% (94 runs sampled)
sample null x 115,404,510 ops/sec ±0.70% (95 runs sampled)
sample [] x 84,197,168 ops/sec ±0.79% (90 runs sampled)
lodash.sample x 44,758,830 ops/sec ±0.59% (94 runs sampled)
lodash.sample null x 50,702,929 ops/sec ±0.55% (91 runs sampled)
lodash.sample [] x 54,129,884 ops/sec ±0.56% (89 runs sampled)
```