@jdalton 2017-02-20
Hi @DenisCarriere!

I'm not so concerned with the nullish case.